pyproject.toml
[project]
name = "sema4-demo"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "braintrust>=0.0.172",
    "autoevals",
    "openai>=1.0",
    "modal>=1.3.2",
    "python-dotenv",
]
